Added Bindings to the PreliminarySurveyPage. Bugs with PreliminarySurveyPage eradicated. Deleted unnecesary usings
