+++ OncoFitness/App.xaml
@@ -9,24 +9,95 @@
         <ResourceDictionary>
             <Color x:Key="Primary">#2196F3</Color>
             <Style TargetType="Button">
-                <Setter Property="TextColor" Value="White"></Setter>
+                <Setter Property="TextColor" 
+                        Value="White">
+                </Setter>
                 <Setter Property="VisualStateManager.VisualStateGroups">
                     <VisualStateGroupList>
                         <VisualStateGroup x:Name="CommonStates">
                             <VisualState x:Name="Normal">
                                 <VisualState.Setters>
-                                    <Setter Property="BackgroundColor" Value="{StaticResource Primary}" />
+                                    <Setter Property="BackgroundColor" 
+                                            Value="{StaticResource Primary}" />
                                 </VisualState.Setters>
                             </VisualState>
                             <VisualState x:Name="Disabled">
                                 <VisualState.Setters>
-                                    <Setter Property="BackgroundColor" Value="#332196F3" />
+                                    <Setter Property="BackgroundColor" 
+                                            Value="#332196F3" />
                                 </VisualState.Setters>
                             </VisualState>
                         </VisualStateGroup>
                     </VisualStateGroupList>
                 </Setter>
             </Style>
+            
+            <ControlTemplate x:Key="RadioButtonTemplate">
+                <Frame BorderColor="#F3F2F1" 
+                       BackgroundColor="#F3F2F1" 
+                       HasShadow="False" 
+                       HeightRequest="70" 
+                       WidthRequest="60" 
+                       HorizontalOptions="Start" 
+                       VerticalOptions="Start" 
+                       Padding="0">
+
+                    <VisualStateManager.VisualStateGroups>
+                        <VisualStateGroupList>
+                            <VisualStateGroup x:Name="CheckedStates">
+
+                                <VisualState x:Name="Checked">
+                                    <VisualState.Setters>
+                                        <Setter Property="BorderColor" 
+                                                Value="DeepPink"/>
+                                        <Setter TargetName="Check" 
+                                                Property="Opacity" 
+                                                Value="1"/>
+                                    </VisualState.Setters>
+                                </VisualState>
+
+                                <VisualState x:Name="Unchecked">
+                                    <VisualState.Setters>
+                                        <Setter Property="BackgroundColor" 
+                                                Value="#f3f2f1"/>
+                                        <Setter Property="BorderColor" 
+                                                Value="#f3f2f1"/>
+                                        <Setter TargetName="Check" 
+                                                Property="Opacity" 
+                                                Value="0"/>
+                                    </VisualState.Setters>
+                                </VisualState>
+
+                            </VisualStateGroup>
+                        </VisualStateGroupList>
+                    </VisualStateManager.VisualStateGroups>
+
+                    <Grid Margin="4" 
+                          WidthRequest="60">
+                        <Grid WidthRequest="12" 
+                              HeightRequest="12" 
+                              HorizontalOptions="End" 
+                              VerticalOptions="Start">
+                            <Ellipse Stroke="DimGray"
+                                     WidthRequest="12" 
+                                     HeightRequest="12" 
+                                     StrokeThickness="0.5" 
+                                     VerticalOptions="Center" 
+                                     HorizontalOptions="Center" 
+                                     Fill="White">
+                            </Ellipse>
+                            <Ellipse x:Name="Check" 
+                                     WidthRequest="6" 
+                                     HeightRequest="6" 
+                                     Fill="DeepPink"  
+                                     VerticalOptions="Center" 
+                                     HorizontalOptions="Center" />
+                        </Grid>
+                        <ContentPresenter></ContentPresenter>
+                    </Grid>
+                </Frame>
+            </ControlTemplate>
+            
         </ResourceDictionary>        
     </Application.Resources>
 </Application>
