+++ OncoFitness/Views/AboutPage.xaml
@@ -37,7 +37,7 @@
                 <Label Text="Успешных тренировок!" FontSize="16" Padding="0,0,0,0" HorizontalTextAlignment="Center"/>
 
                 <Button Margin="0,10,0,0" Text="Начать заниматься"
-                        Command="{Binding OpenWebCommand}"
+                        Command="{Binding StartSurveyCommand}"
                         BackgroundColor="{StaticResource Primary}"
                         TextColor="White" />
