+++ OncoFitness/Views/HistoryPage.xaml.cs
@@ -1,10 +1,4 @@
 ﻿using OncoFitness.ViewModels;
-using System;
-using System.Collections.ObjectModel;
-using System.ComponentModel;
-using System.Linq;
-using System.Threading.Tasks;
-
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
