+++ OncoFitness/Views/NewPlanPage.xaml.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using OncoFitness.ViewModels;
+﻿using OncoFitness.ViewModels;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
